feat: enable processor faults
The function enable_processor_faults() enables the processor faults and
all handlers were created. The processor faults are important as the
next commits will deal directly with the processor core therefore it is
very likelly that processor faults are generated.

The code inside the main function tests the fault generation with an
undef instruction on address 0x20010000. The code was verified with the
debugger and generated the UsageFault